+++ code/renderers/react/src/__test__/Button.stories.tsx
@@ -1,13 +1,12 @@
-import React, { useState } from 'react';
-import { within, userEvent, fn, expect } from '@storybook/test';
-import type { StoryFn as CSF2Story, StoryObj as CSF3Story, Meta } from '..';
+import React, { useEffect, useState } from 'react';
+import { expect, fn, mocked, userEvent, within } from '@storybook/test';
+import type { Meta, StoryFn as CSF2Story, StoryObj as CSF3Story } from '..';
 
 import type { ButtonProps } from './Button';
 import { Button } from './Button';
 import type { HandlerFunction } from '@storybook/addon-actions';
 import { action } from '@storybook/addon-actions';
-import { mocked } from '@storybook/test';
-import { act } from 'react-dom/test-utils';
+import { createPortal } from 'react-dom';
 
 const meta = {
   title: 'Example/Button',
@@ -210,5 +209,66 @@ export const WithActionArgType: CSF3Story<{ someActionArg: HandlerFunction }> = 
   },
   render: () => {
     return <div>nothing</div>;
+  },
+};
+
+export const Modal: CSF3Story = {
+  render: function Component() {
+    const [isModalOpen, setIsModalOpen] = useState(false);
+    const [modalContainer] = useState(() => {
+      const div = document.createElement('div');
+      div.id = 'modal-root';
+      return div;
+    });
+
+    useEffect(() => {
+      document.body.appendChild(modalContainer);
+      return () => {
+        document.body.removeChild(modalContainer);
+      };
+    }, [modalContainer]);
+
+    const handleOpenModal = () => setIsModalOpen(true);
+    const handleCloseModal = () => setIsModalOpen(false);
+
+    const modalContent = isModalOpen
+      ? createPortal(
+          <div
+            role="dialog"
+            style={{
+              position: 'fixed',
+              top: '20%',
+              left: '50%',
+              transform: 'translate(-50%, -20%)',
+              backgroundColor: 'white',
+              padding: '20px',
+              zIndex: 1000,
+              border: '2px solid black',
+              borderRadius: '5px',
+            }}
+          >
+            <div style={{ marginBottom: '10px' }}>
+              <p>This is a modal!</p>
+            </div>
+            <button onClick={handleCloseModal}>Close</button>
+          </div>,
+          modalContainer
+        )
+      : null;
+
+    return (
+      <>
+        <button id="openModalButton" onClick={handleOpenModal}>
+          Open Modal
+        </button>
+        {modalContent}
+      </>
+    );
+  },
+  play: async ({ canvasElement }) => {
+    const canvas = within(canvasElement);
+    const openModalButton = await canvas.getByRole('button', { name: /open modal/i });
+    await userEvent.click(openModalButton);
+    await expect(within(document.body).getByRole('dialog')).toBeInTheDocument();
   },
 };
