+++ code/renderers/react/src/__test__/portable-stories-legacy.test.tsx
@@ -2,15 +2,15 @@
 
 /* eslint-disable import/namespace */
 import React from 'react';
-import { vi, it, expect, afterEach, describe } from 'vitest';
-import { render, screen, cleanup } from '@testing-library/react';
+import { afterEach, describe, expect, it, vi } from 'vitest';
+import { cleanup, render, screen } from '@testing-library/react';
 import { addons } from '@storybook/preview-api';
 
 import * as addonActionsPreview from '@storybook/addon-actions/preview';
 import type { Meta } from '@storybook/react';
 import { expectTypeOf } from 'expect-type';
 
-import { setProjectAnnotations, composeStories, composeStory } from '..';
+import { composeStories, composeStory, setProjectAnnotations } from '..';
 import type { Button } from './Button';
 import * as stories from './Button.stories';
 
@@ -192,6 +192,7 @@ describe('Legacy Portable Stories API', () => {
   const testCases = Object.values(composeStories(stories)).map(
     (Story) => [Story.storyName, Story] as [string, typeof Story]
   );
+
   it.each(testCases)('Renders %s story', async (_storyName, Story) => {
     cleanup();
 
@@ -203,7 +204,14 @@ describe('Legacy Portable Stories API', () => {
 
     const { baseElement } = await render(<Story />);
 
+    globalThis.IS_REACT_ACT_ENVIRONMENT = false;
     await Story.play?.();
+    globalThis.IS_REACT_ACT_ENVIRONMENT = true;
+
     expect(baseElement).toMatchSnapshot();
   });
 });
+
+declare const globalThis: {
+  IS_REACT_ACT_ENVIRONMENT?: boolean;
+};
